fix: refactor accordion
