+++ src/components/common/accordion/Accordion.tsx
@@ -1,31 +1,15 @@
 import React from 'react';
 import classnames from 'classnames';
-import { AccordionProps, BodyProps, HeaderProps } from './types';
+import { Body } from './accordion-body/AccordionBody';
+import { Header } from './accordion-header/AccordionHeader';
 
-const Header: React.FC<HeaderProps> = ({ children, ...props }) => {
-  const randomId = Math.random().toString(36);
-  const { className, id = randomId } = props;
-  const classNames = classnames('accordion-header', className);
-
-  return (
-    <React.Fragment>
-      <input type="checkbox" id={id} hidden />
-      <label className={classNames} htmlFor={id}>
-        {children}
-      </label>
-    </React.Fragment>
-  );
+type Props = {
+  className?: string;
 };
-
-const Body: React.FC<BodyProps> = ({ children, ...props }) => {
-  const { className, ...otherProps } = props;
-  const classNames = classnames('accordion-body', className);
-  return (
-    <div className={classNames} {...otherProps}>
-      {children}
-    </div>
-  );
-};
+interface AccordionFC extends React.FC<React.PropsWithChildren<Props>> {
+  Body: typeof Body;
+  Header: typeof Header;
+}
 
 const Accordion: AccordionFC = ({ children, ...props }) => {
   const { className, ...otherProps } = props;
@@ -36,11 +20,6 @@ const Accordion: AccordionFC = ({ children, ...props }) => {
     </div>
   );
 };
-
-interface AccordionFC extends React.FC<AccordionProps> {
-  Body: typeof Body;
-  Header: typeof Header;
-}
 
 Accordion.Body = Body;
 Accordion.Header = Header;
